Move DemoFeatures to DevModel
