--- engine/demo/src/main/scala/pl/touk/nussknacker/engine/demo/service/SimpleTypesService.scala
@@ -1,46 +1,0 @@
-package pl.touk.nussknacker.engine.demo.service
-
-import javax.annotation.Nullable
-import pl.touk.nussknacker.engine.api.editor._
-import pl.touk.nussknacker.engine.api.{MethodToInvoke, ParamName, Service}
-
-import scala.concurrent.Future
-
-class SimpleTypesService extends Service with Serializable {
-  @MethodToInvoke
-  def invoke(@ParamName("Registered")
-             @SimpleEditor(
-               `type` = SimpleEditorType.BOOL_EDITOR
-             ) registered: Boolean,
-
-             @ParamName("Email")
-             @DualEditor(
-               simpleEditor = new SimpleEditor(`type` = SimpleEditorType.STRING_EDITOR),
-               defaultMode = DualEditorMode.SIMPLE
-             ) email: String,
-
-             @ParamName("BusinessConfig")
-             @DualEditor(
-               simpleEditor = new SimpleEditor(`type` = SimpleEditorType.STRING_EDITOR),
-               defaultMode = DualEditorMode.SIMPLE
-             )
-             @Nullable businessConfig: String,
-
-             @ParamName("Age")
-             @RawEditor
-             @Nullable
-             age: Int,
-
-             @ParamName("CampaignType")
-             @SimpleEditor(
-               `type` = SimpleEditorType.FIXED_VALUES_EDITOR,
-               possibleValues = Array(
-                 new LabeledExpression(expression = "mail", label = "Mail campaign"),
-                 new LabeledExpression(expression = "sms", label = "Sms campaign"),
-                 new LabeledExpression(expression = "popup", label = "Popup campaign"),
-                 new LabeledExpression(expression = "push", label = "Push campaign")
-               )
-             ) campaignType: String
-            ): Future[Unit.type]
-  = Future.successful(Unit)
-}
