+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/dict/RGBDictionary.scala
@@ -1,12 +1,12 @@
-package pl.touk.nussknacker.engine.demo.dicts
+package pl.touk.nussknacker.engine.management.sample.dict
 
 import pl.touk.nussknacker.engine.api.dict.DictInstance
 import pl.touk.nussknacker.engine.api.dict.embedded.EmbeddedDictDefinition
 
-object DictionaryRGB {
-  val dictId: String = "rgb"
+object RGBDictionary {
+  val id: String = "rgb"
 
-  val dictDefinition: EmbeddedDictDefinition = EmbeddedDictDefinition(Map(
+  val definition: EmbeddedDictDefinition = EmbeddedDictDefinition(Map(
     "#000000" -> "Black",
     "#800000" -> "Maroon",
     "#008000" -> "Green",
@@ -19,5 +19,5 @@ object DictionaryRGB {
     "#ffff00" -> "Yellow"
   ))
 
-  val dictInstance: DictInstance = DictInstance(dictId, dictDefinition)
+  val instance: DictInstance = DictInstance(id, definition)
 }
