+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/dict/TestDictionary.scala
@@ -1,0 +1,16 @@
+package pl.touk.nussknacker.engine.management.sample.dict
+
+import pl.touk.nussknacker.engine.api.dict.DictInstance
+import pl.touk.nussknacker.engine.api.dict.embedded.EmbeddedDictDefinition
+
+object TestDictionary {
+  val id: String = "dict"
+
+  val definition: EmbeddedDictDefinition = EmbeddedDictDefinition(Map(
+    "foo" -> "Foo",
+    "bar" -> "Bar",
+    "sentence-with-spaces-and-dots" -> "Sentence with spaces and . dots"
+  ))
+
+  val instance: DictInstance = DictInstance(id, definition)
+}
