+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/dto/Client.scala
@@ -1,0 +1,6 @@
+package pl.touk.nussknacker.engine.management.sample.dto
+
+import io.circe.generic.JsonCodec
+import pl.touk.nussknacker.engine.api.DisplayJsonWithEncoder
+
+@JsonCodec case class Client(id: String, name: String) extends DisplayJsonWithEncoder[Client]
