+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/dto/ConstantState.scala
@@ -1,0 +1,5 @@
+package pl.touk.nussknacker.engine.management.sample.dto
+
+import io.circe.generic.JsonCodec
+
+@JsonCodec case class ConstantState(id: String, transactionId: Int, elements: List[String])
