+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/dto/RichObject.scala
@@ -1,0 +1,3 @@
+package pl.touk.nussknacker.engine.management.sample.dto
+
+case class RichObject(field1: String, field2: Long, field3: Option[String])
