+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/service/SimpleTypesCustomStreamTransformer.scala
@@ -1,0 +1,15 @@
+package pl.touk.nussknacker.engine.management.sample.service
+
+import pl.touk.nussknacker.engine.api.{CustomStreamTransformer, LazyParameter, MethodToInvoke, ParamName}
+
+// In custom stream by default all parameters are eagerly evaluated, you need to define type LazyParameter to make it lazy
+class SimpleTypesCustomStreamTransformer extends CustomStreamTransformer with Serializable {
+  @MethodToInvoke(returnType = classOf[Void])
+  def invoke(@ParamName("booleanParam") booleanParam: Boolean,
+             @ParamName("lazyBooleanParam") lazyBooleanParam: LazyParameter[Boolean],
+             @ParamName("stringParam") string: String,
+             @ParamName("intParam") intParam: Int,
+             @ParamName("bigDecimalParam") bigDecimalParam: java.math.BigDecimal,
+             @ParamName("bigIntegerParam") bigIntegerParam: java.math.BigInteger): Unit = {
+  }
+}
