+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/source/BoundedSource.scala
@@ -1,0 +1,18 @@
+package pl.touk.nussknacker.engine.management.sample.source
+
+import org.apache.flink.streaming.api.scala.StreamExecutionEnvironment
+import pl.touk.nussknacker.engine.api.{MethodToInvoke, ParamName}
+import pl.touk.nussknacker.engine.api.typed.typing.Unknown
+import pl.touk.nussknacker.engine.flink.api.process.FlinkSourceFactory
+import pl.touk.nussknacker.engine.flink.util.source.CollectionSource
+
+import scala.collection.JavaConverters._
+import org.apache.flink.streaming.api.scala._
+
+object BoundedSource extends FlinkSourceFactory[Any] {
+
+  @MethodToInvoke
+  def source(@ParamName("elements") elements: java.util.List[Any]) =
+    new CollectionSource[Any](StreamExecutionEnvironment.getExecutionEnvironment.getConfig, elements.asScala.toList, None, Unknown)
+
+}
