+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/transformer/CustomFilter.scala
@@ -1,0 +1,17 @@
+package pl.touk.nussknacker.engine.management.sample.transformer
+
+import org.apache.flink.streaming.api.scala.DataStream
+import pl.touk.nussknacker.engine.api.{Context, CustomStreamTransformer, LazyParameter, MethodToInvoke, ParamName, ValueWithContext}
+import pl.touk.nussknacker.engine.flink.api.process.{FlinkCustomNodeContext, FlinkCustomStreamTransformation}
+import org.apache.flink.streaming.api.scala._
+
+case object CustomFilter extends CustomStreamTransformer {
+
+  @MethodToInvoke(returnType = classOf[Void])
+  def execute(@ParamName("expression") expression: LazyParameter[Boolean]): FlinkCustomStreamTransformation
+  = FlinkCustomStreamTransformation((start: DataStream[Context], ctx: FlinkCustomNodeContext) =>
+    start
+      .filter(ctx.lazyParameterHelper.lazyFilterFunction(expression))
+      .map(ValueWithContext[Any](null, _)))
+
+}
