+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/transformer/NoneReturnTypeTransformer.scala
@@ -1,0 +1,8 @@
+package pl.touk.nussknacker.engine.management.sample.transformer
+
+import pl.touk.nussknacker.engine.api.{CustomStreamTransformer, LazyParameter, MethodToInvoke, ParamName}
+
+case object NoneReturnTypeTransformer extends CustomStreamTransformer {
+  @MethodToInvoke(returnType = classOf[Void])
+  def execute(@ParamName("expression") expression: LazyParameter[Boolean]): Unit = {}
+}
