+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/transformer/StatefulTransformer.scala
@@ -1,0 +1,28 @@
+package pl.touk.nussknacker.engine.management.sample.transformer
+
+import com.typesafe.scalalogging.LazyLogging
+import org.apache.flink.streaming.api.scala.DataStream
+import pl.touk.nussknacker.engine.api.{Context, CustomStreamTransformer, LazyParameter, MethodToInvoke, ParamName, ValueWithContext}
+import pl.touk.nussknacker.engine.flink.api.process.{FlinkCustomNodeContext, FlinkCustomStreamTransformation}
+import org.apache.flink.streaming.api.scala._
+
+case object StatefulTransformer extends CustomStreamTransformer with LazyLogging {
+
+  @MethodToInvoke
+  def execute(@ParamName("keyBy") keyBy: LazyParameter[String]): FlinkCustomStreamTransformation
+  = FlinkCustomStreamTransformation((start: DataStream[Context], ctx: FlinkCustomNodeContext) => {
+    start
+      .map(ctx.lazyParameterHelper.lazyMapFunction(keyBy))
+      .keyBy(_.value)
+      .mapWithState[ValueWithContext[Any], List[String]] { case (StringFromIr(ir, sr), oldState) =>
+        logger.info(s"received: $sr, current state: $oldState")
+        val nList = sr :: oldState.getOrElse(Nil)
+        (ValueWithContext(nList, ir.context), Some(nList))
+      }
+  })
+
+  object StringFromIr {
+    def unapply(ir: ValueWithContext[_]): Option[(ValueWithContext[_], String)] = Some(ir, ir.context.apply[String]("input"))
+  }
+
+}
