+++ nussknacker-dist/src/universal/conf/dev-application.conf
@@ -10,6 +10,7 @@ categoriesConfig: {
   "Default": "streaming-generic"
   "Category1": "streaming"
   "Category2": "streaming"
+  "DemoFeatures": "streaming"
   "ServerRestApi": "request-response"
 }
